fix: discovery response cache not being utilised
the original design intended to cache the Discovery response to conform hateoas
https://github.com/anomaly/gallagher/issues/5

overriding the pydantic method model_post_init to set the good_known_since var
introduced an error where the test first test would run the subsequent calls would fail
https://docs.pydantic.dev/2.0/api/main/\#pydantic.main.BaseModel.model_post_init

debugging this i found that the discover was running every time as the good_known_since
would never have been set and the cache check would fail.

initially thought of this as a bug with how parsing was working but turns out that if
the cache was detected then we were returning without initialising the configuration
against the discovery response.