refactor: drop pdf field cache
i've just realised that the customer detail response sends the relevant
pdf fields as part of the personalDataDefinitions field, if we can
parse this and make sense of it then the keys to the dynamic fields will
exists in the response and we should not read this from the cache.

moreover it would make sense that the personDataDefinitions is the set
of keys that are part of the customer response i.e not every field is
available or is in use all the time.

changing tactic here, and going to try and parse the
personalDatadefinitions first and then populate the keys from there

refs #1